Updated Notebooks 03 and 04 (#319)
* Updated notebooks to remove one-variable-per-gauge

Removed documentation in the tutorials about 1-variable-per-gauge as it was causing problems with Temp_max and Temp_min being separated and not being able to calculate Temp_ave, whereas having all variables in the same netcdf file did allow so. To prevent further similar issues and let Raven parse correctly, this has been modified in notebooks 03 and 04, i.e. when generating the files and when using them to emulate hydrological models

* Updated NB04 to have the pre-run version

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>