[DO NOT MERGE] Add proposed temporary RavenPy logos (#428)
### What kind of change does this PR introduce?

* This is a PR to work on and discuss the potential RavenPy Logo

### Does this PR introduce a breaking change?

No.

### Other information:

When merging the final version of this PR, we should use "Squash and
Merge" in order to not save lots of images in our git history.

The temporary logos were made by grouping colours via a filter and doing
some light touch-ups in GIMP, then extracting the bitmap and rescaling
the SVG using Inkscape. There's probably a better way to keep some
shading, but this seemed easy and reduces the size of the images to
something more reasonable.

Ideally, we would want a small (150x150) and large (300x300) version of
the logo that we land on in PNG format.

---------

Signed-off-by: Trevor James Smith <10819524+Zeitsperre@users.noreply.github.com>
Co-authored-by: lou-a <louise.arnal@hotmail.fr>